✨ feat(stock_handler): 优化黄金价格处理逻辑
- 添加新的utility包用于处理时区
- 重构黄金价格处理逻辑，使用新的AddPointSeries方法
- 优化时间解析和过滤逻辑
- 更新图表标题格式

♻️ refactor(vadvisor): 重构MultiSeriesLineGraph类型

- 将Numeric接口拆分为ValidType和Numeric
- 将AxesStruct的泛型参数移除，统一使用float64
- 添加AddPointSeries方法用于批量添加数据点
- 优化SetRange方法，支持自动计算范围

🐛 fix(trend_handler): 修复SetRange类型转换问题

- 将SetRange参数从int转换为float64

+++ handler/larkhandler/lark_command/handlers/trend_handler.go
@@ -86,7 +86,7 @@ func TrendHandler(ctx context.Context, data *larkim.P2MessageReceiveV1, args ...
 	title := fmt.Sprintf("[%s]水群频率表-%ddays", larkutils.GetChatName(ctx, *data.Event.Message.ChatId), days)
 	graph.
 		SetTitle(title).
-		SetRange(*min, *max)
+		SetRange(float64(*min), float64(*max))
 	err = larkutils.ReplyCardTextGraph(
 		ctx,
 		"",
